requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
fastapi
Flask
frozenlist
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
httpx-sse
idna
itsdangerous
Jinja2
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langsmith
MarkupSafe
marshmallow
multidict
mypy-extensions
numpy
orjson
packaging
pip
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydantic-settings
pyparsing
python-dotenv
PyYAML
requests
requests-toolbelt
rsa
sniffio
soupsieve
SQLAlchemy
starlette
tenacity
tqdm
typing_extensions
typing-inspect
uritemplate
urllib3
uvicorn
Werkzeug
yarl